+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 ### Added
 
+- Additional file globs to which global configuration should be applied can now
+  be specified on the command line with `--fallback-globs`. Without this option,
+  global configuration is only applied to languages Helix knows about.
+
 ### Fixed
 
 ## [1.0.0] - 2025-01-17